rasa/requirements.txt
rasa==3.3.0
rasa-sdk==3.3.0
faiss-cpu==1.7.3
scikit-learn==0.24.2
numpy==1.21.4
absl-py==0.9.0
SQLAlchemy>=1.4.0,<1.5.0

